+++ inertia/components/captage/chroniques-par-substances.tsx
@@ -1,25 +1,10 @@
 import { useState } from 'react'
 import { fr } from '@codegouvfr/react-dsfr'
 import Select from '@codegouvfr/react-dsfr/Select'
-import ResumeCard from '~/ui/ResumeCard'
-import 'chartjs-adapter-date-fns'
-import {
-  Chart as ChartJS,
-  TimeScale,
-  LinearScale,
-  PointElement,
-  LineElement,
-  Tooltip,
-  Legend,
-  type ChartDataset,
-  type ChartOptions,
-} from 'chart.js'
-import { Chart } from 'react-chartjs-2'
 import Loader from '~/ui/Loader'
 import type { SubstanceItem, ChroniqueData } from '#types/captage'
 import { useFetch } from '~/hooks/use-fetch'
-
-ChartJS.register(TimeScale, LinearScale, PointElement, LineElement, Tooltip, Legend)
+import { SubstanceDetail } from './substance-detail'
 
 type Props = {
   aacCode: string
@@ -27,128 +12,6 @@ type Props = {
   yearMin: number
   yearMax: number
 }
-
-// ─── Scatter chart ────────────────────────────────────────────────────────────
-
-function formatUnite(code: string): string {
-  return code === 'SANS OBJET' ? '' : code
-}
-
-function SubstanceScatterChart({ data }: { data: ChroniqueData }) {
-  const { info, series } = data
-  const unite = formatUnite(info.code_unite)
-
-  const toTs = (d: string) => new Date(d).getTime()
-  const allTs = series.map((p) => toTs(p.date))
-  const minTs = Math.min(...allTs)
-  const maxTs = Math.max(...allTs)
-  // pad bounds slightly so threshold lines extend slightly past first/last point
-  const padMs = (maxTs - minTs) * 0.02
-
-  const conformes = series.filter((p) => p.statut === 'conforme')
-  const depAlerte = series.filter((p) => p.statut === 'dep_alerte')
-  const depRegl = series.filter((p) => p.statut === 'dep_regl')
-
-  const pointDs = (
-    label: string,
-    points: typeof series,
-    color: string
-  ): ChartDataset<'scatter'> => ({
-    type: 'scatter' as const,
-    label,
-    data: points.map((p) => ({ x: toTs(p.date), y: p.valeur })),
-    backgroundColor: color,
-    pointRadius: 5,
-    pointHoverRadius: 7,
-  })
-
-  const lineDs = (label: string, y: number, color: string): ChartDataset<'line'> => ({
-    type: 'line' as const,
-    label,
-    data: [
-      { x: minTs - padMs, y },
-      { x: maxTs + padMs, y },
-    ],
-    borderColor: color,
-    borderDash: [6, 3],
-    borderWidth: 2,
-    pointRadius: 0,
-    fill: false,
-  })
-
-  const datasets: ChartDataset<'scatter' | 'line'>[] = [
-    pointDs('Conforme', conformes, '#003189'),
-    pointDs('Au dessus du seuil réglementaire', depRegl, '#D40000'),
-    pointDs("Au dessus du seuil d'alerte", depAlerte, '#E18B00'),
-    ...(info.seuil_regl !== null
-      ? [lineDs(`Seuil réglementaire (${info.seuil_regl} ${unite})`, info.seuil_regl, '#D40000')]
-      : []),
-    ...(info.seuil_alerte !== null
-      ? [lineDs(`Seuil d'alerte (${info.seuil_alerte} ${unite})`, info.seuil_alerte, '#E18B00')]
-      : []),
-  ]
-
-  const options: ChartOptions = {
-    responsive: true,
-    maintainAspectRatio: false,
-    scales: {
-      x: {
-        type: 'time',
-        time: { unit: 'year', displayFormats: { year: 'yyyy' } },
-        title: { display: false },
-      },
-      y: {
-        beginAtZero: true,
-        title: {
-          display: true,
-          text: `Concentration (${unite})`,
-          font: { size: 12 },
-        },
-      },
-    },
-    plugins: {
-      legend: {
-        display: true,
-        position: 'bottom' as const,
-        align: 'start' as const,
-        labels: { boxWidth: 12, boxHeight: 12, font: { size: 12 } },
-      },
-      tooltip: {
-        backgroundColor: '#fff',
-        titleColor: '#000',
-        bodyColor: '#000',
-        borderColor: 'rgba(0,0,0,0.1)',
-        borderWidth: 1,
-        padding: 10,
-        callbacks: {
-          title: (items) => {
-            const ts = items[0]?.parsed?.x
-            if (!ts) return ''
-            return new Date(ts).toLocaleDateString('fr-FR', {
-              day: 'numeric',
-              month: 'long',
-              year: 'numeric',
-            })
-          },
-          label: (ctx) => {
-            const dsLabel = ctx.dataset.label ?? ''
-            const y = ctx.parsed?.y
-            if (y === undefined) return dsLabel
-            return `${dsLabel} : ${y} ${unite}`
-          },
-        },
-      },
-    },
-  }
-
-  return (
-    <div style={{ height: 320 }}>
-      <Chart type="scatter" data={{ datasets: datasets as any }} options={options} />
-    </div>
-  )
-}
-
-// ─── Main component ───────────────────────────────────────────────────────────
 
 export default function ChroniquesParSubstances({
   aacCode,
@@ -221,155 +84,13 @@ export default function ChroniquesParSubstances({
         )}
       </div>
 
-      {/* Detail panels */}
       {loadingChronique && (
         <div className="flex items-center gap-2 fr-py-2w">
           <Loader type="dots" size="sm" />
         </div>
       )}
 
-      {!loadingChronique && chronique && (
-        <div className="grid grid-cols-2 gap-4" style={{ alignItems: 'start' }}>
-          {/* Left: Informations générales */}
-          <div
-            className="flex flex-col gap-3 fr-p-3w"
-            style={{
-              border: `1px solid ${fr.colors.decisions.border.default.grey.default}`,
-            }}
-          >
-            <h6 className="fr-text--md fr-mb-0" style={{ fontWeight: 700 }}>
-              <span className="fr-icon-list-unordered fr-icon--sm fr-mr-1v" aria-hidden="true" />
-              Informations générales
-            </h6>
-
-            <div className="flex flex-col gap-1">
-              <p className="fr-text--sm fr-mb-0">
-                <span style={{ color: fr.colors.decisions.text.mention.grey.default }}>
-                  Code SANDRE :{' '}
-                </span>
-                <strong>{chronique.info.code_parametre}</strong>
-              </p>
-              {formatUnite(chronique.info.code_unite) && (
-                <p className="fr-text--sm fr-mb-0">
-                  <span style={{ color: fr.colors.decisions.text.mention.grey.default }}>
-                    Unité :{' '}
-                  </span>
-                  <strong>{formatUnite(chronique.info.code_unite)}</strong>
-                </p>
-              )}
-            </div>
-
-            {(chronique.info.seuil_regl !== null || chronique.info.seuil_alerte !== null) && (
-              <div className="flex flex-col gap-1">
-                <hr className="fr-hr fr-mt-1w fr-mb-1w" />
-                <p
-                  className="fr-text--md fr-mb-1v"
-                  style={{
-                    fontWeight: 700,
-                    color: fr.colors.decisions.text.mention.grey.default,
-                  }}
-                >
-                  Seuils
-                </p>
-                {chronique.info.seuil_regl !== null && (
-                  <p className="fr-text--sm fr-mb-0">
-                    <span style={{ color: fr.colors.decisions.text.mention.grey.default }}>
-                      Seuil réglementaire :{' '}
-                    </span>
-                    <strong>
-                      {chronique.info.seuil_regl} {formatUnite(chronique.info.code_unite)}
-                    </strong>
-                  </p>
-                )}
-                {chronique.info.seuil_alerte !== null && (
-                  <p className="fr-text--sm fr-mb-0">
-                    <span style={{ color: fr.colors.decisions.text.mention.grey.default }}>
-                      Seuil d'alerte :{' '}
-                    </span>
-                    <strong>
-                      {chronique.info.seuil_alerte} {formatUnite(chronique.info.code_unite)}
-                    </strong>
-                  </p>
-                )}
-              </div>
-            )}
-
-            {/* Dépassement box */}
-            <ResumeCard
-              size="sm"
-              title="Dépassement réglementaire"
-              value={`${chronique.stats.frequence_dep_regl.toLocaleString('fr-FR')}%`}
-              iconId="fr-icon-warning-line"
-              color={
-                chronique.stats.nb_dep_regl > 0
-                  ? fr.colors.decisions.text.default.error.default
-                  : fr.colors.decisions.text.default.grey.default
-              }
-            />
-          </div>
-
-          {/* Right: Résultats + chart */}
-          <div
-            className="flex flex-col gap-3 fr-p-2w"
-            style={{ border: `1px solid ${fr.colors.decisions.border.default.grey.default}` }}
-          >
-            <h6 className="fr-text--md fr-mb-0" style={{ fontWeight: 700 }}>
-              <span className="fr-icon-microscope-line fr-icon--sm fr-mr-1v" aria-hidden="true" />
-              Résultat des prélèvements sur la période
-            </h6>
-
-            <div className="grid grid-cols-2 gap-2">
-              <ResumeCard
-                size="sm"
-                title="Moyenne de concentration"
-                value={`${chronique.stats.moyenne.toLocaleString('fr-FR')} ${formatUnite(chronique.info.code_unite)}`}
-                iconId="fr-icon-line-chart-line"
-                color={fr.colors.decisions.text.label.blueFrance.default}
-              />
-              <ResumeCard
-                size="sm"
-                title="Maximum des concentrations observés"
-                value={`${chronique.stats.maximum.toLocaleString('fr-FR')} ${formatUnite(chronique.info.code_unite)}`}
-                iconId="fr-icon-warning-line"
-                color={fr.colors.decisions.text.default.warning.default}
-              />
-              <ResumeCard
-                size="sm"
-                title="Résultats en dépassement"
-                value={chronique.stats.nb_dep_regl.toLocaleString('fr-FR')}
-                iconId="fr-icon-error-line"
-                color={
-                  chronique.stats.nb_dep_regl > 0
-                    ? fr.colors.decisions.text.default.error.default
-                    : fr.colors.decisions.text.default.grey.default
-                }
-              />
-              <ResumeCard
-                size="sm"
-                title="Fréquence de dépassement"
-                value={`${chronique.stats.frequence_dep_regl.toLocaleString('fr-FR')}%`}
-                iconId="fr-icon-error-line"
-                color={
-                  chronique.stats.frequence_dep_regl > 0
-                    ? fr.colors.decisions.text.default.error.default
-                    : fr.colors.decisions.text.default.grey.default
-                }
-              />
-            </div>
-
-            {chronique.series.length > 0 ? (
-              <SubstanceScatterChart data={chronique} />
-            ) : (
-              <p
-                className="fr-text--sm fr-mb-0"
-                style={{ color: fr.colors.decisions.text.mention.grey.default }}
-              >
-                Aucune mesure disponible pour cette période.
-              </p>
-            )}
-          </div>
-        </div>
-      )}
+      {!loadingChronique && chronique && <SubstanceDetail chronique={chronique} />}
     </div>
   )
 }
